+++ tests/test_validate_item_collection.py
@@ -1,8 +1,7 @@
 """
-Description: Test the validator
+Description: Test stac-validator on item-collection validation.
 
 """
-
 
 from stac_validator import stac_validator